Authentication Testing
Adding updated authentication files for discussion. Please note, these files will not load properly on GitHub, as we do not maintain an active connection to the server/database. Please reference the domain/test directory on the live site for previewing these files. This is the first step to the dynamic dashboard for organization administrators. 

*Please do not commit to main for now*